Moves KVS-related state store code out into its own set of files.
